+++ Utilidades/utilidades/protocol/unpackers.c
@@ -9,8 +9,8 @@
 void* unpack_SOLICITUD_JOB(int socket){
     payload_SOLICITUD_JOB *payload= malloc(sizeof(payload_SOLICITUD_JOB));
 
-    recv(socket,&(payload->tamanio_nombreArchivo),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreArchivo = payload->tamanio_nombreArchivo;
+    recv(socket,&(payload->tamanio_nombreArchivo),sizeof(int),0);
+    int  tamanio_nombreArchivo = payload->tamanio_nombreArchivo;
 
     char* nombreArchivo = malloc(tamanio_nombreArchivo);
     recv(socket,nombreArchivo,tamanio_nombreArchivo,0);
@@ -22,12 +22,12 @@ void* unpack_SOLICITUD_JOB(int socket){
 void* unpack_ORDEN_TRANSFORMACION(int socket){
     payload_ORDEN_TRANSFORMACION *payload= malloc(sizeof(payload_ORDEN_TRANSFORMACION));
 
-    recv(socket,&(payload->bloque),sizeof(uint16_t),0);
+    recv(socket,&(payload->bloque),sizeof(int),0);
 
-    recv(socket,&(payload->bytesocupados),sizeof(uint32_t),0);
+    recv(socket,&(payload->bytesocupados),sizeof(int),0);
 
-    recv(socket,&(payload->tamanio_nombreArchivoTemporal),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreArchivoTemporal = payload->tamanio_nombreArchivoTemporal;
+    recv(socket,&(payload->tamanio_nombreArchivoTemporal),sizeof(int),0);
+    int  tamanio_nombreArchivoTemporal = payload->tamanio_nombreArchivoTemporal;
 
     char* nombreArchivoTemporal = malloc(tamanio_nombreArchivoTemporal);
     recv(socket,nombreArchivoTemporal,tamanio_nombreArchivoTemporal,0);
@@ -39,15 +39,15 @@ void* unpack_ORDEN_TRANSFORMACION(int socket){
 void* unpack_ORDEN_REDUCCIONLOCAL(int socket){
     payload_ORDEN_REDUCCIONLOCAL *payload= malloc(sizeof(payload_ORDEN_REDUCCIONLOCAL));
 
-    recv(socket,&(payload->tamanio_nombreTemporal_Transformacion),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreTemporal_Transformacion = payload->tamanio_nombreTemporal_Transformacion;
+    recv(socket,&(payload->tamanio_nombreTemporal_Transformacion),sizeof(int),0);
+    int  tamanio_nombreTemporal_Transformacion = payload->tamanio_nombreTemporal_Transformacion;
 
     char* nombreTemporal_Transformacion = malloc(tamanio_nombreTemporal_Transformacion);
     recv(socket,nombreTemporal_Transformacion,tamanio_nombreTemporal_Transformacion,0);
     payload->nombreTemporal_Transformacion = nombreTemporal_Transformacion;
 
-    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionLocal),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreTemporal_ReduccionLocal = payload->tamanio_nombreTemporal_ReduccionLocal;
+    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionLocal),sizeof(int),0);
+    int  tamanio_nombreTemporal_ReduccionLocal = payload->tamanio_nombreTemporal_ReduccionLocal;
 
     char* nombreTemporal_ReduccionLocal = malloc(tamanio_nombreTemporal_ReduccionLocal);
     recv(socket,nombreTemporal_ReduccionLocal,tamanio_nombreTemporal_ReduccionLocal,0);
@@ -59,30 +59,30 @@ void* unpack_ORDEN_REDUCCIONLOCAL(int socket){
 void* unpack_ORDEN_REDUCCIONGLOBAL(int socket){
     payload_ORDEN_REDUCCIONGLOBAL *payload= malloc(sizeof(payload_ORDEN_REDUCCIONGLOBAL));
 
-    recv(socket,&(payload->PUERTO_Nodo),sizeof(uint16_t),0);
+    recv(socket,&(payload->PUERTO_Nodo),sizeof(int),0);
 
-    recv(socket,&(payload->tamanio_IP_Nodo),sizeof(uint16_t),0);
-    uint16_t  tamanio_IP_Nodo = payload->tamanio_IP_Nodo;
+    recv(socket,&(payload->tamanio_IP_Nodo),sizeof(int),0);
+    int  tamanio_IP_Nodo = payload->tamanio_IP_Nodo;
 
     char* IP_Nodo = malloc(tamanio_IP_Nodo);
     recv(socket,IP_Nodo,tamanio_IP_Nodo,0);
     payload->IP_Nodo = IP_Nodo;
 
-    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionLocal),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreTemporal_ReduccionLocal = payload->tamanio_nombreTemporal_ReduccionLocal;
+    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionLocal),sizeof(int),0);
+    int  tamanio_nombreTemporal_ReduccionLocal = payload->tamanio_nombreTemporal_ReduccionLocal;
 
     char* nombreTemporal_ReduccionLocal = malloc(tamanio_nombreTemporal_ReduccionLocal);
     recv(socket,nombreTemporal_ReduccionLocal,tamanio_nombreTemporal_ReduccionLocal,0);
     payload->nombreTemporal_ReduccionLocal = nombreTemporal_ReduccionLocal;
 
-    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionGlobal),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreTemporal_ReduccionGlobal = payload->tamanio_nombreTemporal_ReduccionGlobal;
+    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionGlobal),sizeof(int),0);
+    int  tamanio_nombreTemporal_ReduccionGlobal = payload->tamanio_nombreTemporal_ReduccionGlobal;
 
     char* nombreTemporal_ReduccionGlobal = malloc(tamanio_nombreTemporal_ReduccionGlobal);
     recv(socket,nombreTemporal_ReduccionGlobal,tamanio_nombreTemporal_ReduccionGlobal,0);
     payload->nombreTemporal_ReduccionGlobal = nombreTemporal_ReduccionGlobal;
 
-    recv(socket,&(payload->encargado),sizeof(uint16_t),0);
+    recv(socket,&(payload->encargado),sizeof(int),0);
 
     return (void*)payload;
 };
@@ -90,8 +90,8 @@ void* unpack_ORDEN_REDUCCIONGLOBAL(int socket){
 void* unpack_ORDEN_ALMACENAMIENTO(int socket){
     payload_ORDEN_ALMACENAMIENTO *payload= malloc(sizeof(payload_ORDEN_ALMACENAMIENTO));
 
-    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionGlobal),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreTemporal_ReduccionGlobal = payload->tamanio_nombreTemporal_ReduccionGlobal;
+    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionGlobal),sizeof(int),0);
+    int  tamanio_nombreTemporal_ReduccionGlobal = payload->tamanio_nombreTemporal_ReduccionGlobal;
 
     char* nombreTemporal_ReduccionGlobal = malloc(tamanio_nombreTemporal_ReduccionGlobal);
     recv(socket,nombreTemporal_ReduccionGlobal,tamanio_nombreTemporal_ReduccionGlobal,0);
@@ -103,21 +103,21 @@ void* unpack_ORDEN_ALMACENAMIENTO(int socket){
 void* unpack_INFO_TRANSFORMACION(int socket){
     payload_INFO_TRANSFORMACION *payload= malloc(sizeof(payload_INFO_TRANSFORMACION));
 
-    recv(socket,&(payload->PUERTO_Worker),sizeof(uint16_t),0);
+    recv(socket,&(payload->PUERTO_Worker),sizeof(int),0);
 
-    recv(socket,&(payload->tamanio_IP_Worker),sizeof(uint16_t),0);
-    uint16_t  tamanio_IP_Worker = payload->tamanio_IP_Worker;
+    recv(socket,&(payload->tamanio_IP_Worker),sizeof(int),0);
+    int  tamanio_IP_Worker = payload->tamanio_IP_Worker;
 
     char* IP_Worker = malloc(tamanio_IP_Worker);
     recv(socket,IP_Worker,tamanio_IP_Worker,0);
     payload->IP_Worker = IP_Worker;
 
-    recv(socket,&(payload->bloque),sizeof(uint16_t),0);
+    recv(socket,&(payload->bloque),sizeof(int),0);
 
-    recv(socket,&(payload->bytesocupados),sizeof(uint32_t),0);
+    recv(socket,&(payload->bytesocupados),sizeof(int),0);
 
-    recv(socket,&(payload->tamanio_nombreArchivoTemporal),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreArchivoTemporal = payload->tamanio_nombreArchivoTemporal;
+    recv(socket,&(payload->tamanio_nombreArchivoTemporal),sizeof(int),0);
+    int  tamanio_nombreArchivoTemporal = payload->tamanio_nombreArchivoTemporal;
 
     char* nombreArchivoTemporal = malloc(tamanio_nombreArchivoTemporal);
     recv(socket,nombreArchivoTemporal,tamanio_nombreArchivoTemporal,0);
@@ -129,24 +129,24 @@ void* unpack_INFO_TRANSFORMACION(int socket){
 void* unpack_INFO_REDUCCIONLOCAL(int socket){
     payload_INFO_REDUCCIONLOCAL *payload= malloc(sizeof(payload_INFO_REDUCCIONLOCAL));
 
-    recv(socket,&(payload->PUERTO_Worker),sizeof(uint16_t),0);
+    recv(socket,&(payload->PUERTO_Worker),sizeof(int),0);
 
-    recv(socket,&(payload->tamanio_IP_Worker),sizeof(uint16_t),0);
-    uint16_t  tamanio_IP_Worker = payload->tamanio_IP_Worker;
+    recv(socket,&(payload->tamanio_IP_Worker),sizeof(int),0);
+    int  tamanio_IP_Worker = payload->tamanio_IP_Worker;
 
     char* IP_Worker = malloc(tamanio_IP_Worker);
     recv(socket,IP_Worker,tamanio_IP_Worker,0);
     payload->IP_Worker = IP_Worker;
 
-    recv(socket,&(payload->tamanio_nombreTemporal_Transformacion),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreTemporal_Transformacion = payload->tamanio_nombreTemporal_Transformacion;
+    recv(socket,&(payload->tamanio_nombreTemporal_Transformacion),sizeof(int),0);
+    int  tamanio_nombreTemporal_Transformacion = payload->tamanio_nombreTemporal_Transformacion;
 
     char* nombreTemporal_Transformacion = malloc(tamanio_nombreTemporal_Transformacion);
     recv(socket,nombreTemporal_Transformacion,tamanio_nombreTemporal_Transformacion,0);
     payload->nombreTemporal_Transformacion = nombreTemporal_Transformacion;
 
-    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionLocal),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreTemporal_ReduccionLocal = payload->tamanio_nombreTemporal_ReduccionLocal;
+    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionLocal),sizeof(int),0);
+    int  tamanio_nombreTemporal_ReduccionLocal = payload->tamanio_nombreTemporal_ReduccionLocal;
 
     char* nombreTemporal_ReduccionLocal = malloc(tamanio_nombreTemporal_ReduccionLocal);
     recv(socket,nombreTemporal_ReduccionLocal,tamanio_nombreTemporal_ReduccionLocal,0);
@@ -158,30 +158,30 @@ void* unpack_INFO_REDUCCIONLOCAL(int socket){
 void* unpack_INFO_REDUCCIONGLOBAL(int socket){
     payload_INFO_REDUCCIONGLOBAL *payload= malloc(sizeof(payload_INFO_REDUCCIONGLOBAL));
 
-    recv(socket,&(payload->PUERTO_Worker),sizeof(uint16_t),0);
+    recv(socket,&(payload->PUERTO_Worker),sizeof(int),0);
 
-    recv(socket,&(payload->tamanio_IP_Worker),sizeof(uint16_t),0);
-    uint16_t  tamanio_IP_Worker = payload->tamanio_IP_Worker;
+    recv(socket,&(payload->tamanio_IP_Worker),sizeof(int),0);
+    int  tamanio_IP_Worker = payload->tamanio_IP_Worker;
 
     char* IP_Worker = malloc(tamanio_IP_Worker);
     recv(socket,IP_Worker,tamanio_IP_Worker,0);
     payload->IP_Worker = IP_Worker;
 
-    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionLocal),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreTemporal_ReduccionLocal = payload->tamanio_nombreTemporal_ReduccionLocal;
+    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionLocal),sizeof(int),0);
+    int  tamanio_nombreTemporal_ReduccionLocal = payload->tamanio_nombreTemporal_ReduccionLocal;
 
     char* nombreTemporal_ReduccionLocal = malloc(tamanio_nombreTemporal_ReduccionLocal);
     recv(socket,nombreTemporal_ReduccionLocal,tamanio_nombreTemporal_ReduccionLocal,0);
     payload->nombreTemporal_ReduccionLocal = nombreTemporal_ReduccionLocal;
 
-    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionGlobal),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreTemporal_ReduccionGlobal = payload->tamanio_nombreTemporal_ReduccionGlobal;
+    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionGlobal),sizeof(int),0);
+    int  tamanio_nombreTemporal_ReduccionGlobal = payload->tamanio_nombreTemporal_ReduccionGlobal;
 
     char* nombreTemporal_ReduccionGlobal = malloc(tamanio_nombreTemporal_ReduccionGlobal);
     recv(socket,nombreTemporal_ReduccionGlobal,tamanio_nombreTemporal_ReduccionGlobal,0);
     payload->nombreTemporal_ReduccionGlobal = nombreTemporal_ReduccionGlobal;
 
-    recv(socket,&(payload->encargado),sizeof(uint16_t),0);
+    recv(socket,&(payload->encargado),sizeof(int),0);
 
     return (void*)payload;
 };
@@ -189,17 +189,17 @@ void* unpack_INFO_REDUCCIONGLOBAL(int socket){
 void* unpack_INFO_ALMACENAMIENTO(int socket){
     payload_INFO_ALMACENAMIENTO *payload= malloc(sizeof(payload_INFO_ALMACENAMIENTO));
 
-    recv(socket,&(payload->PUERTO_Worker),sizeof(uint16_t),0);
+    recv(socket,&(payload->PUERTO_Worker),sizeof(int),0);
 
-    recv(socket,&(payload->tamanio_IP_Worker),sizeof(uint16_t),0);
-    uint16_t  tamanio_IP_Worker = payload->tamanio_IP_Worker;
+    recv(socket,&(payload->tamanio_IP_Worker),sizeof(int),0);
+    int  tamanio_IP_Worker = payload->tamanio_IP_Worker;
 
     char* IP_Worker = malloc(tamanio_IP_Worker);
     recv(socket,IP_Worker,tamanio_IP_Worker,0);
     payload->IP_Worker = IP_Worker;
 
-    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionGlobal),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreTemporal_ReduccionGlobal = payload->tamanio_nombreTemporal_ReduccionGlobal;
+    recv(socket,&(payload->tamanio_nombreTemporal_ReduccionGlobal),sizeof(int),0);
+    int  tamanio_nombreTemporal_ReduccionGlobal = payload->tamanio_nombreTemporal_ReduccionGlobal;
 
     char* nombreTemporal_ReduccionGlobal = malloc(tamanio_nombreTemporal_ReduccionGlobal);
     recv(socket,nombreTemporal_ReduccionGlobal,tamanio_nombreTemporal_ReduccionGlobal,0);
@@ -211,8 +211,8 @@ void* unpack_INFO_ALMACENAMIENTO(int socket){
 void* unpack_PETICION_NODO(int socket){
     payload_PETICION_NODO *payload= malloc(sizeof(payload_PETICION_NODO));
 
-    recv(socket,&(payload->tamanio_nombreArchivo),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreArchivo = payload->tamanio_nombreArchivo;
+    recv(socket,&(payload->tamanio_nombreArchivo),sizeof(int),0);
+    int  tamanio_nombreArchivo = payload->tamanio_nombreArchivo;
 
     char* nombreArchivo = malloc(tamanio_nombreArchivo);
     recv(socket,nombreArchivo,tamanio_nombreArchivo,0);
@@ -224,17 +224,17 @@ void* unpack_PETICION_NODO(int socket){
 void* unpack_NODO(int socket){
     payload_NODO *payload= malloc(sizeof(payload_NODO));
 
-    recv(socket,&(payload->PUERTO_Nodo),sizeof(uint16_t),0);
+    recv(socket,&(payload->PUERTO_Nodo),sizeof(int),0);
 
-    recv(socket,&(payload->tamanio_IP_Nodo),sizeof(uint16_t),0);
-    uint16_t  tamanio_IP_Nodo = payload->tamanio_IP_Nodo;
+    recv(socket,&(payload->tamanio_IP_Nodo),sizeof(int),0);
+    int  tamanio_IP_Nodo = payload->tamanio_IP_Nodo;
 
     char* IP_Nodo = malloc(tamanio_IP_Nodo);
     recv(socket,IP_Nodo,tamanio_IP_Nodo,0);
     payload->IP_Nodo = IP_Nodo;
 
-    recv(socket,&(payload->tamanio_nombreNodo),sizeof(uint16_t),0);
-    uint16_t  tamanio_nombreNodo = payload->tamanio_nombreNodo;
+    recv(socket,&(payload->tamanio_nombreNodo),sizeof(int),0);
+    int  tamanio_nombreNodo = payload->tamanio_nombreNodo;
 
     char* nombreNodo = malloc(tamanio_nombreNodo);
     recv(socket,nombreNodo,tamanio_nombreNodo,0);
@@ -272,11 +272,11 @@ void* unpack_BLOQUE(int socket){
 void* unpack_PRESENTACION_DATANODE(int socket){
     payload_PRESENTACION_DATANODE *payload= malloc(sizeof(payload_PRESENTACION_DATANODE));
 
-    recv(socket,&(payload->pid),sizeof(uint16_t),0);
+    recv(socket,&(payload->pid),sizeof(int),0);
 
-    recv(socket,&(payload->id_dataNode),sizeof(uint16_t),0);
+    recv(socket,&(payload->id_dataNode),sizeof(int),0);
 
-    recv(socket,&(payload->cantidad_bloques),sizeof(uint16_t),0);
+    recv(socket,&(payload->cantidad_bloques),sizeof(int),0);
 
     return (void*)payload;
 };
@@ -284,7 +284,7 @@ void* unpack_PRESENTACION_DATANODE(int socket){
 void* unpack_JOB(int socket){
     payload_JOB *payload= malloc(sizeof(payload_JOB));
 
-    recv(socket,&(payload->id_master),sizeof(uint16_t),0);
+    recv(socket,&(payload->id_master),sizeof(int),0);
 
     return (void*)payload;
 };
@@ -292,13 +292,13 @@ void* unpack_JOB(int socket){
 void* unpack_RESPUESTA_MASTER(int socket){
     payload_RESPUESTA_MASTER *payload= malloc(sizeof(payload_RESPUESTA_MASTER));
 
-    recv(socket,&(payload->id_master),sizeof(uint16_t),0);
+    recv(socket,&(payload->id_master),sizeof(int),0);
 
-    recv(socket,&(payload->id_nodo),sizeof(uint16_t),0);
+    recv(socket,&(payload->id_nodo),sizeof(int),0);
 
-    recv(socket,&(payload->bloque),sizeof(uint16_t),0);
+    recv(socket,&(payload->bloque),sizeof(int),0);
 
-    recv(socket,&(payload->estado),sizeof(uint16_t),0);
+    recv(socket,&(payload->estado),sizeof(int),0);
 
     return (void*)payload;
 };
@@ -306,8 +306,8 @@ void* unpack_RESPUESTA_MASTER(int socket){
 void* unpack_SCRIPT(int socket){
     payload_SCRIPT *payload= malloc(sizeof(payload_SCRIPT));
 
-    recv(socket,&(payload->tamanio_contenido),sizeof(uint16_t),0);
-    uint16_t  tamanio_contenido = payload->tamanio_contenido;
+    recv(socket,&(payload->tamanio_contenido),sizeof(int),0);
+    int  tamanio_contenido = payload->tamanio_contenido;
 
     char* contenido = malloc(tamanio_contenido);
     recv(socket,contenido,tamanio_contenido,0);
@@ -319,7 +319,7 @@ void* unpack_SCRIPT(int socket){
 void* unpack_PETICION_BLOQUE(int socket){
     payload_PETICION_BLOQUE *payload= malloc(sizeof(payload_PETICION_BLOQUE));
 
-    recv(socket,&(payload->numero_bloque),sizeof(uint32_t),0);
+    recv(socket,&(payload->numero_bloque),sizeof(int),0);
 
     return (void*)payload;
 };
